+++ Nymph/src/main/kotlin/com/ambrosia/nymph/services/ItemService.kt
@@ -28,8 +28,9 @@ class ItemService(
 
     @Transactional
     fun editItem(businessId: Long, itemId: Long, itemDto: ItemDto): ItemDto {
-        businessRepository.findById(businessId)
-            .orElseThrow { EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId)) }
+        if (!businessRepository.existsById(businessId)) {
+            throw EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId))
+        }
         val item = itemRepository.findById(itemId)
             .orElseThrow { EntityNotFoundException(Item::class.java, mutableMapOf("id" to itemId)) }
         itemDto.name?.let { item.name = it }
@@ -43,8 +44,9 @@ class ItemService(
 
     @Transactional
     fun deleteItem(businessId: Long, itemId: Long) {
-        businessRepository.findById(businessId)
-            .orElseThrow { EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId)) }
+        if (!businessRepository.existsById(businessId)) {
+            throw EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId))
+        }
         val item = itemRepository.findById(itemId)
             .orElseThrow { EntityNotFoundException(Item::class.java, mutableMapOf("id" to itemId)) }
         itemRepository.delete(item)
@@ -52,8 +54,9 @@ class ItemService(
 
     @Transactional
     fun getItems(businessId: Long): List<ItemDto> {
-        businessRepository.findById(businessId)
-            .orElseThrow { EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId)) }
+        if (!businessRepository.existsById(businessId)) {
+            throw EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId))
+        }
         return itemRepository.findByBusinessId(businessId).stream().map { it.toDto() }.toList()
     }
 }
